[feature] Support download model from ModelScope
